+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # SCSS-Lint Changelog
 
-## master (unreleased)
+## 0.28.0
 
 * Fix bug in `StringQuotes` where strings with interpolation were incorrectly
   being linted